Move tests out.
